Isolated file CRUD operations to file_utils.rs and reimplemented them into main.
